requirements.txt
Flask==2.1.0
gunicorn==20.1.0
PyYAML==6.0